ch07/myproject_virtualenv/src/django-myproject/requirements/_base.txt
asgiref==3.2.3
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.3
defusedxml==0.6.0
Django==3.0rc1
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-imagekit==4.0.2
ecdsa==0.14.1
future==0.18.2
idna==2.8
oauthlib==3.1.0
pilkit==2.0
Pillow==6.2.1
psycopg2-binary==2.8.3
pyasn1==0.4.8
PyJWT==1.7.1
python-dotenv==0.10.3
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
